Clearly separate Blueprints v1 code paths from Blueprints v2 code paths
