+++ packages/playground/cli/src/blueprints-v2/blueprint-v2-declaration.ts
@@ -1,0 +1,45 @@
+import type { BlueprintDeclaration } from '@wp-playground/blueprints';
+
+export type BlueprintV2Declaration = string | BlueprintDeclaration | undefined;
+export type ParsedBlueprintV2Declaration =
+	| { type: 'inline-file'; contents: string }
+	| { type: 'file-reference'; reference: string };
+
+export function parseBlueprintDeclaration(
+	source: BlueprintV2Declaration | ParsedBlueprintV2Declaration
+): ParsedBlueprintV2Declaration {
+	if (
+		typeof source === 'object' &&
+		'type' in source &&
+		['inline-file', 'file-reference'].includes(source.type)
+	) {
+		return source;
+	}
+	if (!source) {
+		return {
+			type: 'inline-file',
+			contents: '{}',
+		};
+	}
+	if (typeof source !== 'string') {
+		// If source is an object, assume it's a Blueprint declaration object and
+		// convert it to a JSON string.
+		return {
+			type: 'inline-file',
+			contents: JSON.stringify(source),
+		};
+	}
+	try {
+		// If source is valid JSON, return it as is.
+		JSON.parse(source);
+		return {
+			type: 'inline-file',
+			contents: source,
+		};
+	} catch {
+		return {
+			type: 'file-reference',
+			reference: source,
+		};
+	}
+}
