+++ packages/playground/cli/src/load-balancer.ts
@@ -1,6 +1,6 @@
 import type { PHPRequest, PHPResponse, RemoteAPI } from '@php-wasm/universal';
-import type { PlaygroundCliBlueprintV1Worker as PlaygroundCliWorkerV1 } from './worker-thread';
-import type { PlaygroundCliBlueprintV2Worker as PlaygroundCliWorkerV2 } from './worker-thread-v2';
+import type { PlaygroundCliBlueprintV1Worker as PlaygroundCliWorkerV1 } from './blueprints-v1/worker-thread-v1';
+import type { PlaygroundCliBlueprintV2Worker as PlaygroundCliWorkerV2 } from './blueprints-v2/worker-thread-v2';
 
 type PlaygroundCliWorker = PlaygroundCliWorkerV1 | PlaygroundCliWorkerV2;
