+++ packages/playground/cli/src/run-cli.ts
@@ -1,25 +1,15 @@
 import { errorLogPath, logger } from '@php-wasm/logger';
-import { EmscriptenDownloadMonitor, ProgressTracker } from '@php-wasm/progress';
 import type {
 	PHPRequest,
 	RemoteAPI,
 	SupportedPHPVersion,
 } from '@php-wasm/universal';
-import {
-	PHPResponse,
-	consumeAPI,
-	exposeAPI,
-	exposeSyncAPI,
-} from '@php-wasm/universal';
+import { PHPResponse, exposeAPI, exposeSyncAPI } from '@php-wasm/universal';
 import type {
 	BlueprintBundle,
 	BlueprintDeclaration,
 } from '@wp-playground/blueprints';
-import {
-	compileBlueprint,
-	isBlueprintBundle,
-	runBlueprintSteps,
-} from '@wp-playground/blueprints';
+import { runBlueprintSteps } from '@wp-playground/blueprints';
 import {
 	RecommendedPHPVersion,
 	unzipFile,
@@ -27,33 +17,19 @@ import {
 } from '@wp-playground/common';
 import fs from 'fs';
 import type { Server } from 'http';
-import path from 'path';
 import { MessageChannel as NodeMessageChannel, Worker } from 'worker_threads';
 // @ts-ignore
-import { resolveWordPressRelease } from '@wp-playground/wordpress';
-import {
-	CACHE_FOLDER,
-	cachedDownload,
-	fetchSqliteIntegration,
-	readAsFile,
-} from './download';
 import {
 	expandAutoMounts,
 	parseMountDirArguments,
 	parseMountWithDelimiterArguments,
 } from './mounts';
-import { startServer } from './server';
-import type { Mount, PlaygroundCliBlueprintV1Worker } from './worker-thread';
+import { startServer } from './start-server';
 import type {
-	PlaygroundCliBlueprintV2Worker,
-	WorkerBootArgs,
-} from './worker-thread-v2';
-// @ts-ignore
-import importedWorkerV1UrlString from './worker-thread?worker&url';
-// @ts-ignore
-import importedWorkerV2UrlString from './worker-thread-v2?worker&url';
-
-// @ts-ignore
+	Mount,
+	PlaygroundCliBlueprintV1Worker,
+} from './blueprints-v1/worker-thread-v1';
+import type { PlaygroundCliBlueprintV2Worker } from './blueprints-v2/worker-thread-v2';
 import { FileLockManagerForNode } from '@php-wasm/node';
 import { LoadBalancer } from './load-balancer';
 /* eslint-disable no-console */
@@ -64,6 +40,8 @@ import type { MessagePort as NodeMessagePort } from 'worker_threads';
 import yargs from 'yargs';
 import { isValidWordPressSlug } from './is-valid-wordpress-slug';
 import { resolveBlueprint } from './resolve-blueprint';
+import { BlueprintsV2Handler } from './blueprints-v2/blueprints-v2-handler';
+import { BlueprintsV1Handler } from './blueprints-v1/blueprints-v1-handler';
 
 export async function parseOptionsAndRunCLI() {
 	/**
@@ -564,385 +542,7 @@ export async function runCLI(args: RunCLIArgs): Promise<RunCLIServer> {
 	});
 }
 
-/**
- * Boots Playground CLI workers using Blueprint version 2.
- *
- * Progress tracking, downloads, steps, and all other features are
- * implemented in PHP and orchestrated by the worker thread.
- */
-class BlueprintsV2Handler {
-	private phpVersion: SupportedPHPVersion;
-	private lastProgressMessage = '';
-
-	private siteUrl: string;
-	private processIdSpaceLength: number;
-	private args: RunCLIArgs;
-
-	constructor(
-		args: RunCLIArgs,
-		options: {
-			siteUrl: string;
-			processIdSpaceLength: number;
-		}
-	) {
-		this.args = args;
-		this.siteUrl = options.siteUrl;
-		this.processIdSpaceLength = options.processIdSpaceLength;
-		this.phpVersion = args.php as SupportedPHPVersion;
-	}
-
-	getWorkerUrl() {
-		return importedWorkerV2UrlString;
-	}
-
-	async bootPrimaryWorker(
-		phpPort: NodeMessagePort,
-		fileLockManagerPort: NodeMessagePort
-	) {
-		const playground: RemoteAPI<PlaygroundCliBlueprintV2Worker> =
-			consumeAPI(phpPort);
-
-		await playground.useFileLockManager(fileLockManagerPort);
-
-		const workerBootArgs = {
-			...this.args,
-			php: this.phpVersion,
-			siteUrl: this.siteUrl,
-			firstProcessId: 1,
-			processIdSpaceLength: this.processIdSpaceLength,
-			trace: this.args.debug || false,
-			blueprint: this.args.blueprint!,
-		};
-
-		await playground.bootAsPrimaryWorker(workerBootArgs);
-		return playground;
-	}
-
-	async bootSecondaryWorker({
-		worker,
-		fileLockManagerPort,
-		firstProcessId,
-	}: {
-		worker: SpawnedWorker;
-		fileLockManagerPort: NodeMessagePort;
-		firstProcessId: number;
-	}) {
-		const playground: RemoteAPI<PlaygroundCliBlueprintV2Worker> =
-			consumeAPI(worker.phpPort);
-
-		await playground.useFileLockManager(fileLockManagerPort);
-
-		const workerBootArgs: WorkerBootArgs = {
-			...this.args,
-			php: this.phpVersion!,
-			siteUrl: this.siteUrl,
-			firstProcessId,
-			processIdSpaceLength: this.processIdSpaceLength,
-			trace: this.args.debug || false,
-			blueprint: this.args.blueprint!,
-		};
-
-		await playground.bootAsSecondaryWorker(workerBootArgs);
-
-		return playground;
-	}
-
-	writeProgressUpdate(
-		writeStream: NodeJS.WriteStream,
-		message: string,
-		finalUpdate: boolean
-	) {
-		if (message === this.lastProgressMessage) {
-			// Avoid repeating the same message
-			return;
-		}
-		this.lastProgressMessage = message;
-
-		if (writeStream.isTTY) {
-			// Overwrite previous progress updates in-place for a quieter UX.
-			writeStream.cursorTo(0);
-			writeStream.write(message);
-			writeStream.clearLine(1);
-
-			if (finalUpdate) {
-				writeStream.write('\n');
-			}
-		} else {
-			// Fall back to writing one line per progress update
-			writeStream.write(`${message}\n`);
-		}
-	}
-}
-
-/**
- * Boots Playground CLI workers using Blueprint version 1.
- *
- * Progress tracking, downloads, steps, and all other features are
- * implemented in TypeScript and orchestrated by this class.
- */
-class BlueprintsV1Handler {
-	private phpVersion: SupportedPHPVersion | undefined;
-	private lastProgressMessage = '';
-
-	private siteUrl: string;
-	private processIdSpaceLength: number;
-	private args: RunCLIArgs;
-
-	constructor(
-		args: RunCLIArgs,
-		options: {
-			siteUrl: string;
-			processIdSpaceLength: number;
-		}
-	) {
-		this.args = args;
-		this.siteUrl = options.siteUrl;
-		this.processIdSpaceLength = options.processIdSpaceLength;
-	}
-
-	getWorkerUrl() {
-		return importedWorkerV1UrlString;
-	}
-
-	async bootPrimaryWorker(
-		phpPort: NodeMessagePort,
-		fileLockManagerPort: NodeMessagePort
-	) {
-		const compiledBlueprint = await this.compileInputBlueprint(
-			this.args['additional-blueprint-steps'] || []
-		);
-		this.phpVersion = compiledBlueprint.versions.php;
-
-		let wpDetails: any = undefined;
-		// @TODO: Rename to FetchProgressMonitor. There's nothing Emscripten
-		// about that class anymore.
-		const monitor = new EmscriptenDownloadMonitor();
-		if (!this.args.skipWordPressSetup) {
-			let progressReached100 = false;
-			monitor.addEventListener('progress', ((
-				e: CustomEvent<ProgressEvent & { finished: boolean }>
-			) => {
-				if (progressReached100) {
-					return;
-				}
-
-				// @TODO Every progress bar will want percentages. The
-				//       download monitor should just provide that.
-				const { loaded, total } = e.detail;
-				// Use floor() so we don't report 100% until truly there.
-				const percentProgress = Math.floor(
-					Math.min(100, (100 * loaded) / total)
-				);
-				progressReached100 = percentProgress === 100;
-
-				if (!this.args.quiet) {
-					this.writeProgressUpdate(
-						process.stdout,
-						`Downloading WordPress ${percentProgress}%...`,
-						progressReached100
-					);
-				}
-			}) as any);
-
-			wpDetails = await resolveWordPressRelease(this.args.wp);
-			logger.log(
-				`Resolved WordPress release URL: ${wpDetails?.releaseUrl}`
-			);
-		}
-
-		const preinstalledWpContentPath =
-			wpDetails &&
-			path.join(
-				CACHE_FOLDER,
-				`prebuilt-wp-content-for-wp-${wpDetails.version}.zip`
-			);
-		const wordPressZip = !wpDetails
-			? undefined
-			: fs.existsSync(preinstalledWpContentPath)
-			? readAsFile(preinstalledWpContentPath)
-			: await cachedDownload(
-					wpDetails.releaseUrl,
-					`${wpDetails.version}.zip`,
-					monitor
-			  );
-
-		logger.log(`Fetching SQLite integration plugin...`);
-		const sqliteIntegrationPluginZip = this.args.skipSqliteSetup
-			? undefined
-			: await fetchSqliteIntegration(monitor);
-
-		const followSymlinks = this.args.followSymlinks === true;
-		const trace = this.args.experimentalTrace === true;
-
-		const mountsBeforeWpInstall = this.args['mount-before-install'] || [];
-		const mountsAfterWpInstall = this.args.mount || [];
-
-		const playground = consumeAPI<PlaygroundCliBlueprintV1Worker>(phpPort);
-
-		// Comlink communication proxy
-		await playground.isConnected();
-
-		logger.log(`Booting WordPress...`);
-
-		await playground.useFileLockManager(fileLockManagerPort);
-		await playground.bootAsPrimaryWorker({
-			phpVersion: this.phpVersion,
-			wpVersion: compiledBlueprint.versions.wp,
-			absoluteUrl: this.siteUrl,
-			mountsBeforeWpInstall,
-			mountsAfterWpInstall,
-			wordPressZip: wordPressZip && (await wordPressZip!.arrayBuffer()),
-			sqliteIntegrationPluginZip:
-				await sqliteIntegrationPluginZip!.arrayBuffer(),
-			firstProcessId: 0,
-			processIdSpaceLength: this.processIdSpaceLength,
-			followSymlinks,
-			trace,
-			internalCookieStore: this.args.internalCookieStore,
-			withXdebug: this.args.xdebug,
-		});
-
-		if (
-			wpDetails &&
-			!this.args['mount-before-install'] &&
-			!fs.existsSync(preinstalledWpContentPath)
-		) {
-			logger.log(`Caching preinstalled WordPress for the next boot...`);
-			fs.writeFileSync(
-				preinstalledWpContentPath,
-				(await zipDirectory(playground, '/wordpress'))!
-			);
-			logger.log(`Cached!`);
-		}
-
-		return playground;
-	}
-
-	async bootSecondaryWorker({
-		worker,
-		fileLockManagerPort,
-		firstProcessId,
-	}: {
-		worker: SpawnedWorker;
-		fileLockManagerPort: NodeMessagePort;
-		firstProcessId: number;
-	}) {
-		const additionalPlayground = consumeAPI<PlaygroundCliBlueprintV1Worker>(
-			worker.phpPort
-		);
-
-		await additionalPlayground.isConnected();
-		await additionalPlayground.useFileLockManager(fileLockManagerPort);
-		await additionalPlayground.bootAsSecondaryWorker({
-			phpVersion: this.phpVersion,
-			absoluteUrl: this.siteUrl,
-			mountsBeforeWpInstall: this.args['mount-before-install'] || [],
-			mountsAfterWpInstall: this.args['mount'] || [],
-			// Skip WordPress zip because we share the /wordpress directory
-			// populated by the initial worker.
-			wordPressZip: undefined,
-			// Skip SQLite integration plugin for now because we
-			// will copy it from primary's `/internal` directory.
-			sqliteIntegrationPluginZip: undefined,
-			dataSqlPath: '/wordpress/wp-content/database/.ht.sqlite',
-			firstProcessId,
-			processIdSpaceLength: this.processIdSpaceLength,
-			followSymlinks: this.args.followSymlinks === true,
-			trace: this.args.experimentalTrace === true,
-			// @TODO: Move this to the request handler or else every worker
-			//        will have a separate cookie store.
-			internalCookieStore: this.args.internalCookieStore,
-			withXdebug: this.args.xdebug,
-		});
-		await additionalPlayground.isReady();
-		return additionalPlayground;
-	}
-
-	async compileInputBlueprint(additionalBlueprintSteps: any[]) {
-		const args = this.args;
-		const resolvedBlueprint = args.blueprint as BlueprintDeclaration;
-		/**
-		 * @TODO This looks similar to the resolveBlueprint() call in the website package:
-		 * 	     https://github.com/WordPress/wordpress-playground/blob/ce586059e5885d185376184fdd2f52335cca32b0/packages/playground/website/src/main.tsx#L41
-		 *
-		 * 		 Also the Blueprint Builder tool does something similar.
-		 *       Perhaps all these cases could be handled by the same function?
-		 */
-		const blueprint: BlueprintDeclaration | BlueprintBundle =
-			isBlueprintBundle(resolvedBlueprint)
-				? resolvedBlueprint
-				: {
-						login: args.login,
-						...(resolvedBlueprint || {}),
-						preferredVersions: {
-							php:
-								args.php ??
-								resolvedBlueprint?.preferredVersions?.php ??
-								RecommendedPHPVersion,
-							wp:
-								args.wp ??
-								resolvedBlueprint?.preferredVersions?.wp ??
-								'latest',
-							...(resolvedBlueprint?.preferredVersions || {}),
-						},
-				  };
-
-		const tracker = new ProgressTracker();
-		let lastCaption = '';
-		let progressReached100 = false;
-		tracker.addEventListener('progress', (e: any) => {
-			if (progressReached100) {
-				return;
-			}
-			progressReached100 = e.detail.progress === 100;
-
-			// Use floor() so we don't report 100% until truly there.
-			const progressInteger = Math.floor(e.detail.progress);
-			lastCaption =
-				e.detail.caption || lastCaption || 'Running the Blueprint';
-			const message = `${lastCaption.trim()} – ${progressInteger}%`;
-			if (!args.quiet) {
-				this.writeProgressUpdate(
-					process.stdout,
-					message,
-					progressReached100
-				);
-			}
-		});
-		return await compileBlueprint(blueprint as BlueprintDeclaration, {
-			progress: tracker,
-			additionalSteps: additionalBlueprintSteps,
-		});
-	}
-
-	writeProgressUpdate(
-		writeStream: NodeJS.WriteStream,
-		message: string,
-		finalUpdate: boolean
-	) {
-		if (message === this.lastProgressMessage) {
-			// Avoid repeating the same message
-			return;
-		}
-		this.lastProgressMessage = message;
-
-		if (writeStream.isTTY) {
-			// Overwrite previous progress updates in-place for a quieter UX.
-			writeStream.cursorTo(0);
-			writeStream.write(message);
-			writeStream.clearLine(1);
-
-			if (finalUpdate) {
-				writeStream.write('\n');
-			}
-		} else {
-			// Fall back to writing one line per progress update
-			writeStream.write(`${message}\n`);
-		}
-	}
-}
-
-type SpawnedWorker = {
+export type SpawnedWorker = {
 	worker: Worker;
 	phpPort: NodeMessagePort;
 };
