+++ packages/playground/cli/tests/mount-examples/not-plugin/foo.php
