+++ packages/playground/cli/tests/mount-examples/plugin/sample-plugin.php
