+++ packages/playground/cli/tests/mount-examples/static-html/index.html
