+++ packages/playground/cli/tests/mount-examples/theme/functions.php
