+++ packages/playground/cli/tests/mount-examples/theme/index.php
