+++ packages/playground/cli/tests/mount-examples/wordpress/wp-load.php
