+++ packages/playground/cli/tests/test-running-unbuilt-cli.sh
