+++ packages/playground/cli/tsconfig.spec.json
@@ -6,8 +6,8 @@
 	},
 	"include": [
 		"jest.config.ts",
-		"src/**/*.test.ts",
-		"src/**/*.spec.ts",
-		"src/**/*.d.ts"
+		"tests/**/*.test.ts",
+		"tests/**/*.spec.ts",
+		"tests/**/*.d.ts"
 	]
 }
