+++ packages/playground/cli/vite.config.ts
@@ -135,8 +135,8 @@ export default defineConfig({
 			entry: {
 				index: 'src/index.ts',
 				cli: 'src/cli.ts',
-				'worker-thread': 'src/worker-thread.ts',
-				'worker-thread-v2': 'src/worker-thread-v2.ts',
+				'worker-thread-v1': 'src/blueprints-v1/worker-thread-v1.ts',
+				'worker-thread-v2': 'src/blueprints-v2/worker-thread-v2.ts',
 			},
 			name: 'playground-cli',
 			formats: ['es', 'cjs'],
